DevOps/requirements.txt
# -*- coding: utf-8 -*-
discord.py >= 1.5.0
Jinja2 >= 2.11.3
Pillow >= 8.0.1
PyYAML >= 5.3.1
requests >= 2.25.1
